requirements.txt
fastapi
mysql-connector-python
python-jose
passlib[bcrypt]
python-multipart